+++ README.md
@@ -88,6 +88,8 @@
     - [x] 로또 개수 화면에 출력
     - [x] 생성된 로또를 화면에 출력
   - [x] 당첨번호 및 보너스 번호 입력 창 제작 (css)
+    - # 당첨번호 및 보너스 번호 기능 구현
+    - [x] 당첨번호 및 보너스 번호 가져오기
   - [x] 결과 확인 버튼 제작 (css)
   - # layout
   - [x] header 제작
