+++ index.html
@@ -51,7 +51,7 @@
         <div class="winningLotto-contents body">
           <p>지난 주 당첨번호 6개와 보너스 번호 1개를 입력해주세요.</p>
           <div>
-            <div>
+            <div class="winningLotto-contents_winningLotto">
               <p>당첨 번호</p>
               <div>
                 <input />
